DOC/requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-parler==2.3
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
